+++ source/win/openwin.c
@@ -8,104 +8,18 @@
 #define OPENW_DEF_POSY 1
 
 #define OPENW_WIDTH 80
-#define OPENW_HEIGH 18
+#define OPENW_HEIGH 20
 
 #define MASK_POSX 5
 #define MASK_POSY 3
 
 #define MASK_H 13
-#define MASK_W 75
+#define MASK_W 74
 
 #define MAX_PATHLENGHT_STRING 55
 
 #define WAS_A_FILE 1
 #define WAS_A_DIR 2
-
-void generate_path_str(char* dest){
-    if(strlen(g_current_directory) > MAX_PATHLENGHT_STRING){
-        sprintf(dest + 2, "%s", g_current_directory + strlen(g_current_directory) - MAX_PATHLENGHT_STRING + 2);
-        dest[0] = '.';
-        dest[1] = '.';
-    }
-    else sprintf(dest, "%s", g_current_directory);
-}
-
-int compile_file_list(UserControl uc){
-    if(uc->doc != NULL){
-        dealloc_document(uc->doc);
-    }
-    uc->doc = alloc_document(1, STR_REALLOC_DEF_INTERVAL);
-    if(uc->doc == NULL) return -1; 
-    DIR *d;
-    struct dirent *dir;
-    d = opendir(g_current_directory);
-    if(d == NULL) return -1;
-    errno = 0;
-    while ((dir = readdir(d)) != NULL){
-        if(strcmp(dir->d_name, ".") == 0 || strcmp(dir->d_name, "..") == 0) continue;
-        if(dir->d_type != DT_REG && dir->d_type != DT_DIR) continue;
-        int sz = strlen(dir->d_name);
-        if(dir->d_type == DT_DIR){
-            int result = doc_add_character(uc->doc, '/');
-            if(result < 0) {
-                closedir(d);
-                return result;
-            }
-        }
-        for(int i = 0; i < sz; i++){
-            int result = doc_add_character(uc->doc, (wchar_t)dir->d_name[i]);
-            if(result < 0) {
-                closedir(d);
-                return result;
-            }
-        }
-        int result = doc_cursor_enter(uc->doc);
-        if(result < 0) {
-            closedir(d);
-            return result;
-        }
-    }
-    if(errno != 0) return -1;
-    closedir(d);
-    int result = doc_remove_character(uc->doc);
-    if(result < 0) return result;
-    uc->doc->cursor_x = 0;
-    uc->doc->cursor_y = 0;
-    return 0;
-}
-
-int get_dir_type(char* str){
-    DIR* d;
-    struct dirent *dir;
-    d = opendir(g_current_directory);
-    if(d == NULL) return -1;
-    while ((dir = readdir(d)) != NULL){
-        if(str[0] == '/'){
-            if(strcmp(dir->d_name, str + 1) == 0) { //str is always something like /namefolder, the / is added by the function compile_file_list in the document of this window. Doing +1 ignore that first character
-                closedir(d);
-                return dir->d_type;
-            }
-        }
-        else{
-            if(strcmp(dir->d_name, str) == 0) {
-                closedir(d);
-                return dir->d_type;
-            }
-        }
-    }
-    closedir(d);
-    return -1;
-}
-
-int path_exists(char* str){
-    DIR* d;
-    d = opendir(str);
-    if(d) {
-        closedir(d);
-        return true;
-    }
-    return false;
-}
 
 int on_press_enter(UserControl uc){
     if(uc->doc->size == 1 && uc->doc->lines[0]->_used == 0) return 0; //current folder is empty
@@ -122,7 +36,7 @@ int on_press_enter(UserControl uc){
         //but in I'm in the root folder /, then I don't want to get //newfolder, so this is the reason of this if/else statements
         if(path_exists(new_path)){
             strcpy(g_current_directory, new_path);
-            int result = compile_file_list(uc);
+            int result = compile_file_list(uc->doc);
             if(result < 0) return result;
             return WAS_A_DIR;
         }
@@ -151,8 +65,6 @@ int on_press_ctrlup(UserControl uc){
     if(i > 0) new_dir[i] = '\0';
     if(path_exists(new_dir)){
         strcpy(g_current_directory, new_dir);
-        int result = compile_file_list(uc);
-        if(result <= 0) return result;
         return 0;
     }
     return -1;
@@ -161,7 +73,7 @@ int on_press_ctrlup(UserControl uc){
 int openw_init(UserControl uc){
     WINDOW *savewin = create_window(OPENW_HEIGH, OPENW_WIDTH, OPENW_DEF_POSY, OPENW_DEF_POSX);
     Guiw_mask *mask = alloc_guiw_mask(MASK_H, MASK_W);
-    int result = compile_file_list(uc);
+    int result = compile_file_list(uc->doc);
     if(savewin == NULL || mask == NULL || result < 0) {
         if(mask != NULL) dealloc_guiw_mask(mask);
         return -1;
@@ -182,12 +94,19 @@ int openw_reinit(UserControl uc){
 int openw_draw(UserControl uc){
     update_mask(uc->mask, uc->doc);
     werase(uc->window);
+    box(uc->window, 0, 0);
+    wmove(uc->window, MASK_POSY + MASK_H, 0);
+    wprintw(uc->window, "├────────────────────┄┄┈┈");
+    wmove(uc->window, MASK_POSY + MASK_H + 1, 1);
+    wprintw(uc->window, "Press [CTRL + ↑] or [Page up] to go one folder up");
+    wmove(uc->window, MASK_POSY + MASK_H + 2, 1);
+    wprintw(uc->window, "Press [ENTER] to open the selected file, or go into the selected folder");
     wmove(uc->window, 1, 1);
     wprintw(uc->window, "Open file:");
     wmove(uc->window, 2, 1);
     wprintw(uc->window, "Current directory: ");
     char display_path[MAX_PATHLENGHT_STRING];
-    generate_path_str(display_path);
+    generate_path_str(display_path, MAX_PATHLENGHT_STRING);
     wprintw(uc->window, "%s", display_path);
     for(int i = 0; i < uc->mask->rows; i++){ //print document text from mask matrix
         wmove(uc->window, MASK_POSY + i, MASK_POSX);
@@ -199,7 +118,6 @@ int openw_draw(UserControl uc){
         wmove(uc->window, MASK_POSY, 1);
         wprintw(uc->window, "There is nothing inside this folder");
     }
-    box(uc->window, 0, 0);
     wrefresh(uc->window);
     return 0;
 }
@@ -231,6 +149,8 @@ int openw_handle_input(UserControl uc, wchar_t input, int crm){
                     case CTRL_A_KEY_UP:
                     case KEY_PGUP:
                         on_press_ctrlup(uc);
+                        int result = compile_file_list(uc->doc);
+                        if(result <= 0) return result;
                         break;
                 }
                 reset_cnt_buffer(input_buffer);
